tuturial/dl/model/attention_is_all_you_need/requirements.txt
dill
msgpack-numpy
msgpack-python
torchtext